Move ModuleType and InaccessibleDynamicType to test/ as they are related to typeString parser only
