+++ src/types/ast/internal/index.ts
@@ -1,3 +1,2 @@
 export * from "./function_set";
-export * from "./inaccessible_dynamic_type";
 export * from "./type_var";
