+++ test/utils/typeStrings/ast/inaccessible_dynamic_type.ts
@@ -1,9 +1,8 @@
-import { TypeNode } from "../type";
+import { TypeNode } from "../../../../src";
 
 /**
  * Place-holder type used only by the typestring parser for
  * cases when the typestring contains "inaccessible dynamic type".
- * @todo move under test/ along with typestring parser
  */
 export class InaccessibleDynamicType extends TypeNode {
     pp(): string {
