+++ test/utils/typeStrings/ast/module_type.ts
@@ -1,9 +1,5 @@
-import { Range } from "../../misc";
-import { TypeNode } from "./type";
+import { Range, TypeNode } from "../../../../src";
 
-// NOTE: This class is only used by the TypeString parser and should be
-// considered deprecated.
-// @deprecated
 export class ModuleType extends TypeNode {
     readonly path: string;
